moved bukkit configuration tests from custom package to the correct parent package
This does fix the coverage for junit tests
