+++ src/test/java/org/betonquest/betonquest/api/bukkit/config/custom/fallback/FallbackConfigurationNestedTest.java
@@ -1,8 +1,5 @@
-package org.betonquest.betonquest.api.bukkit.config.custom.fallback.complex;
+package org.betonquest.betonquest.api.bukkit.config.custom.fallback;
 
-import org.betonquest.betonquest.api.bukkit.config.custom.fallback.FallbackConfiguration;
-import org.betonquest.betonquest.api.bukkit.config.custom.fallback.FallbackConfigurationSection;
-import org.betonquest.betonquest.api.bukkit.config.custom.fallback.FallbackConfigurationSectionTest;
 import org.bukkit.configuration.Configuration;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.MemoryConfiguration;
@@ -13,12 +10,12 @@ import java.io.File;
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
- * Tests the {@link FallbackConfigurationSection} class.
+ * Tests the {@link FallbackConfiguration} class.
  */
 @SuppressWarnings("PMD.JUnitAssertionsShouldIncludeMessage")
-public class FallbackConfigurationSectionNestedTest extends FallbackConfigurationSectionTest {
+class FallbackConfigurationNestedTest extends FallbackConfigurationTest {
     @Override
-    public ConfigurationSection getConfig() {
+    public Configuration getConfig() {
         final Configuration original = YamlConfiguration.loadConfiguration(new File("src/test/resources/api/bukkit/fallback/original.yml"));
         fallback = YamlConfiguration.loadConfiguration(new File("src/test/resources/api/bukkit/fallback/fallback.yml"));
